requirements.txt
Flask
pandas
PyMuPDF
google-generativeai
python-dotenv
fpdf2
Werkzeug